Key interaction working as intended.
