+++ src/main/java/com/kireiiiiiiii/shooting_stars/ui/elements/game/HomeButton.java
@@ -1,6 +1,6 @@
 /*
  * Author: Matěj Šťastný
- * Date created: 6/16/2024
+ * Date created: 7/3/2024
  * Github link: https://github.com/kireiiiiiiii/ShootingStars
  *
  *
@@ -24,13 +24,15 @@
  *
  */
 
-package com.kireiiiiiiii.shooting_stars.ui.elements.menu_panel_elements.links_panel;
+package com.kireiiiiiiii.shooting_stars.ui.elements.game;
 
 import java.awt.Graphics2D;
-import java.awt.Container;
 import java.awt.event.MouseEvent;
 import java.util.ArrayList;
+import java.awt.Image;
+import java.awt.Container;
 
+import com.kireiiiiiiii.shooting_stars.constants.Colors;
 import com.kireiiiiiiii.shooting_stars.constants.Interact;
 import com.kireiiiiiiii.shooting_stars.constants.Textures;
 import com.kireiiiiiiii.shooting_stars.constants.WidgetTags;
@@ -40,38 +42,42 @@ import com.kireiiiiiiii.shooting_stars.ui.Interactable;
 import com.kireiiiiiiii.shooting_stars.ui.Renderable;
 
 /**
- * Button that openes my instagram account page in the browser.
+ * Home button widget for the pause screen and the game over screen. Redirects
+ * the user to the main menu.
  * 
  */
-public class InstagramLink implements Renderable, Interactable {
+public class HomeButton implements Renderable, Interactable {
 
     /////////////////
     // Constants
     ////////////////
 
-    private final int[] SIZE = { 90, 90 };
+    private final int[] SIZE = { 70, 70 };
+    private final int[] ICON_PADDING = { 15, 15 };
+    private final int BORDER_WIDTH = 10;
+    private final int ARC_WIDTH = 40;
 
     /////////////////
     // Variables
     ////////////////
 
+    private Image texture;
     private int[] position;
-    private boolean isVisible;
+    private boolean visible;
 
     /////////////////
-    // Constructors
+    // Constructor
     ////////////////
 
     /**
-     * Default constructor.
+     * Default button constructor.
      * 
      * @param owner    - owning {@code JPanel} object.
-     * @param position - position.
+     * @param position - position of the button in the render.
      */
-    public InstagramLink(int[] position) {
+    public HomeButton(int[] position) {
         this.position = position;
-        this.position[0] = this.position[0] + SIZE[0] / 2;
-        this.position[1] = this.position[1] + SIZE[1] / 2;
+        this.texture = Textures.HOME_ICON;
     }
 
     /////////////////
@@ -80,37 +86,50 @@ public class InstagramLink implements Renderable, Interactable {
 
     @Override
     public void render(Graphics2D g, Container img) {
-        if (!this.isVisible) {
+
+        if (!visible) {
             return;
         }
 
-        g.drawImage(ImageUtil.scaleImage(Textures.INSTAGRAM_LOGO, SIZE[0], SIZE[1]), position[0], position[1], img);
+        g.setColor(Colors.MAIN_GRAY);
+        g.fillRoundRect(this.position[0] - BORDER_WIDTH / 2, this.position[1] - BORDER_WIDTH / 2,
+                SIZE[0] + BORDER_WIDTH, SIZE[1] + BORDER_WIDTH, ARC_WIDTH + BORDER_WIDTH, ARC_WIDTH + BORDER_WIDTH);
+        g.setColor(Colors.MAIN_GREEN);
+        g.fillRoundRect(this.position[0], this.position[1], SIZE[0], SIZE[1], ARC_WIDTH, ARC_WIDTH);
+
+        if (this.texture != null) {
+            g.drawImage(
+                    ImageUtil.scaleImage(this.texture, SIZE[0] - ICON_PADDING[0], SIZE[1] - ICON_PADDING[1]),
+                    position[0] + ICON_PADDING[0] / 2, position[1] + ICON_PADDING[1] / 2,
+                    img);
+        }
     }
 
     @Override
     public int getZIndex() {
-        return ZIndexes.POPUP_PANEL_BUTTONS;
+        return ZIndexes.WIDGETS;
     }
 
     @Override
     public boolean isVisible() {
-        return this.isVisible;
+        return this.visible;
     }
 
     @Override
     public void hide() {
-        this.isVisible = false;
+        this.visible = false;
     }
 
     @Override
     public void show() {
-        this.isVisible = true;
+        this.visible = true;
     }
 
     @Override
     public ArrayList<String> getTags() {
         ArrayList<String> tags = new ArrayList<String>();
-        tags.add(WidgetTags.LINKS);
+        tags.add(WidgetTags.PAUSE);
+        tags.add(WidgetTags.GAME_OVER);
         return tags;
     }
 
@@ -120,7 +139,7 @@ public class InstagramLink implements Renderable, Interactable {
 
     @Override
     public Runnable getInteraction() {
-        return Interact.INSTAGRAM;
+        return Interact.HOME;
     }
 
     @Override
