+++ src/main/java/com/kireiiiiiiii/shooting_stars/ui/elements/game/StarWidget.java
@@ -24,7 +24,7 @@
  *
  */
 
-package com.kireiiiiiiii.shooting_stars.ui.elements.game_panel_elements;
+package com.kireiiiiiiii.shooting_stars.ui.elements.game;
 
 import java.awt.Graphics2D;
 import java.awt.Container;
@@ -32,6 +32,7 @@ import java.awt.event.MouseEvent;
 import java.util.ArrayList;
 
 import com.kireiiiiiiii.shooting_stars.constants.Colors;
+import com.kireiiiiiiii.shooting_stars.constants.Interact;
 import com.kireiiiiiiii.shooting_stars.constants.Textures;
 import com.kireiiiiiiii.shooting_stars.constants.WidgetTags;
 import com.kireiiiiiiii.shooting_stars.constants.ZIndexes;
@@ -42,7 +43,7 @@ import com.kireiiiiiiii.shooting_stars.ui.Renderable;
  * The target widget.
  * 
  */
-public class StarWidget implements Renderable {
+public class StarWidget implements Renderable, com.kireiiiiiiii.shooting_stars.ui.Interactable {
 
     /////////////////
     // Constants
@@ -125,22 +126,8 @@ public class StarWidget implements Renderable {
     }
 
     /////////////////
-    // Public methods
+    // Interact
     ////////////////
-
-    /**
-     * Calculates, if the target got clicked depending on its location, and the
-     * location of the click.
-     * 
-     * @param e - {@code MouseEvent} of the interaction.
-     * @return - {@code boolean} if the target was hit.
-     */
-    public boolean wasClicked(MouseEvent e) {
-        int x = e.getX();
-        int y = e.getY();
-        return Math.abs(this.location[0] - x) <= this.radius
-                && Math.abs(this.location[1] - y) <= this.radius;
-    }
 
     /////////////////
     // Modifiers
@@ -162,6 +149,19 @@ public class StarWidget implements Renderable {
      */
     public void setRadius(int radius) {
         this.radius = radius;
+    }
+
+    @Override
+    public boolean wasInteracted(MouseEvent e) {
+        int x = e.getX();
+        int y = e.getY();
+        return Math.abs(this.location[0] - x) <= this.radius
+                && Math.abs(this.location[1] - y) <= this.radius;
+    }
+
+    @Override
+    public Runnable getInteraction() {
+        return Interact.TARGET_INTERACTED;
     }
 
 }
